change event-bus to app-data update app-data to use observable pattern with subscribe, unsubscribe, and next update lessons-counter and lessons-list components to be Observer and subscribe to lessonsList
